// src/types/gfg.ts
export interface GFGUserStats {
  username: string;
  totalSolved: number;
  codingScore: number;
  institute: string;
  instituteRank: number;
  school: number;
  basic: number;
  easy: number;
  medium: number;
  hard: number;
  currentStreak: number;
  maxStreak: number;
  contestRating: number;
}